fix php8.1 compatibility

+++ Service/Processor/SaveTags.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace MageSuite\CmsTagManager\Service\Processor;
 
 class SaveTags
@@ -15,8 +16,7 @@ class SaveTags
     public function __construct(
         \MageSuite\CmsTagManager\Api\TagsRepositoryInterface $tagsRepository,
         \MageSuite\CmsTagManager\Model\TagsFactory $tags
-    )
-    {
+    ) {
         $this->tagsRepository = $tagsRepository;
         $this->tags = $tags;
     }
@@ -26,33 +26,35 @@ class SaveTags
      */
     public function processSave($data)
     {
-        $postTagsArray = explode(',', $data['page_tags']);
+        if (!empty($data['page_tags'])) {
+            $postTagsArray = explode(',', $data['page_tags']);
 
-        $pageTags = $this->tagsRepository->getTagsByCmsPageId($data['page_id']);
+            $pageTags = $this->tagsRepository->getTagsByCmsPageId($data['page_id']);
 
-        $tagsToSkip = [];
-        foreach ($pageTags as $tag) {
-            if (!in_array($tag, $postTagsArray)) {
-                $tag = $this->tagsRepository->getTag($data['page_id'], $tag);
-                $this->tagsRepository->delete($tag);
+            $tagsToSkip = [];
+            foreach ($pageTags as $tag) {
+                if (!in_array($tag, $postTagsArray)) {
+                    $tag = $this->tagsRepository->getTag($data['page_id'], $tag);
+                    $this->tagsRepository->delete($tag);
+                }
+
+                $tagsToSkip[] = $tag;
             }
 
-            $tagsToSkip[] = $tag;
-        }
+            foreach ($postTagsArray as $postTag) {
+                if (!$postTag) {
+                    continue;
+                }
+                if (in_array($postTag, $tagsToSkip)) {
+                    continue;
+                }
 
-        foreach ($postTagsArray as $postTag) {
-            if (!$postTag) {
-                continue;
+                $newTag = $this->tags->create();
+
+                $newTag->setCmsPageId($data['page_id']);
+                $newTag->setTagName($postTag);
+                $this->tagsRepository->save($newTag);
             }
-            if (in_array($postTag, $tagsToSkip)) {
-                continue;
-            }
-
-            $newTag = $this->tags->create();
-
-            $newTag->setCmsPageId($data['page_id']);
-            $newTag->setTagName($postTag);
-            $this->tagsRepository->save($newTag);
         }
     }
 }
